Module lyon_geom::cubic_to_quadratic

Functions

cubic_to_monotonic_quadratics

cubic_to_quadratics

Approximates a cubic bézier segment with a sequence of quadratic béziers.

make_monotonic

single_curve_approximation

This is terrible as a general approximation but works if the cubic curve does not have inflection points and is “flat” enough. Typically usables after subdiving the curve a few times.

single_curve_approximation_error

Evaluates an upper bound on the maximum distance between the curve and its quadratic approximation obtained using the single curve approximation.

single_curve_approximation_test