Module lyon_geom::flatten_cubic

Structs

Flattened
An iterator over a cubic bezier segment that yields line segments approximating the curve for a given approximation threshold.

Functions

find_cubic_bezier_inflection_points

flatten_cubic_bezier

flatten_cubic_no_inflection

flatten_including_inflection

inflection_approximation_range

no_inflection_flattening_step