Struct lyon_geom::quadratic_bezier::QuadraticBezierSegment

pub struct QuadraticBezierSegment<S> {
    pub from: Point<S>,
    pub ctrl: Point<S>,
    pub to: Point<S>,
}

A 2d curve segment defined by three points: the beginning of the segment, a control point and the end of the segment.

The curve is defined by equation: ∀ t ∈ [0..1], P(t) = (1 - t)² * from + 2 * (1 - t) * t * ctrl + t² * to

Fields
from: Point<S>ctrl: Point<S>to: Point<S>
Implementations

impl<S: Scalar> QuadraticBezierSegment<S>

pub fn sample(&self, t: S) -> Point<S>

Sample the curve at t (expecting t between 0 and 1).

pub fn x(&self, t: S) -> S

Sample the x coordinate of the curve at t (expecting t between 0 and 1).

pub fn y(&self, t: S) -> S

Sample the y coordinate of the curve at t (expecting t between 0 and 1).

pub fn derivative(&self, t: S) -> Vector<S>

Sample the curve’s derivative at t (expecting t between 0 and 1).

pub fn dx(&self, t: S) -> S

Sample the x coordinate of the curve’s derivative at t (expecting t between 0 and 1).

pub fn dy(&self, t: S) -> S

Sample the y coordinate of the curve’s derivative at t (expecting t between 0 and 1).

pub fn flip(&self) -> Self

Swap the beginning and the end of the segment.

pub fn y_maximum_t(&self) -> S

Find the advancement of the y-most position in the curve.

This returns the advancement along the curve, not the actual y position.

pub fn y_minimum_t(&self) -> S

Find the advancement of the y-least position in the curve.

This returns the advancement along the curve, not the actual y position.

pub fn local_y_extremum_t(&self) -> Option<S>

Return the y inflection point or None if this curve is y-monotonic.

pub fn x_maximum_t(&self) -> S

Find the advancement of the x-most position in the curve.

This returns the advancement along the curve, not the actual x position.

pub fn x_minimum_t(&self) -> S

Find the advancement of the x-least position in the curve.

This returns the advancement along the curve, not the actual x position.

pub fn local_x_extremum_t(&self) -> Option<S>

Return the x inflection point or None if this curve is x-monotonic.

pub fn split_range(&self, t_range: Range<S>) -> Self

Return the sub-curve inside a given range of t.

This is equivalent splitting at the range’s end points.

pub fn split( &self, t: S, ) -> (QuadraticBezierSegment<S>, QuadraticBezierSegment<S>)

Split this curve into two sub-curves.

pub fn before_split(&self, t: S) -> QuadraticBezierSegment<S>

Return the curve before the split point.

pub fn after_split(&self, t: S) -> QuadraticBezierSegment<S>

Return the curve after the split point.

pub fn to_cubic(&self) -> CubicBezierSegment<S>

Elevate this curve to a third order bézier.

pub fn baseline(&self) -> LineSegment<S>

pub fn is_a_point(&self, tolerance: S) -> bool

Returns whether the curve can be approximated with a single point, given a tolerance threshold.

pub fn is_linear(&self, tolerance: S) -> bool

Returns true if the curve can be approximated with a single line segment given a tolerance threshold.

pub fn fat_line(&self) -> (LineEquation<S>, LineEquation<S>)

Computes a “fat line” of this segment.

A fat line is two conservative lines between which the segment is fully contained.

pub fn transformed<T: Transformation<S>>(&self, transform: &T) -> Self

Applies the transform to this curve and returns the results.

pub fn flattening_step(&self, tolerance: S) -> S

Find the interval of the beginning of the curve that can be approximated with a line segment.

pub fn for_each_flattened<F>(&self, tolerance: S, callback: &mut F)
where F: FnMut(&LineSegment<S>),

Approximates the curve with sequence of line segments.

The tolerance parameter defines the maximum distance between the curve and its approximation.

This implements the algorithm described by Raph Levien at https://raphlinus.github.io/graphics/curves/2019/12/23/flatten-quadbez.html

pub fn for_each_flattened_with_t<F>(&self, tolerance: S, callback: &mut F)
where F: FnMut(&LineSegment<S>, Range<S>),

Compute a flattened approximation of the curve, invoking a callback at each step.

The tolerance parameter defines the maximum distance between the curve and its approximation.

The end of the t parameter range at the final segment is guaranteed to be equal to 1.0.

This implements the algorithm described by Raph Levien at https://raphlinus.github.io/graphics/curves/2019/12/23/flatten-quadbez.html

pub fn flattened(&self, tolerance: S) -> Flattened<S>

Returns the flattened representation of the curve as an iterator, starting after the current point.

pub fn flattened_t(&self, tolerance: S) -> FlattenedT<S>

Returns the flattened representation of the curve as an iterator, starting after the current point.

pub fn for_each_monotonic_range<F>(&self, cb: &mut F)
where F: FnMut(Range<S>),

Invokes a callback for each monotonic part of the segment.

pub fn for_each_monotonic<F>(&self, cb: &mut F)
where F: FnMut(&QuadraticBezierSegment<S>),

Invokes a callback for each monotonic part of the segment.

pub fn for_each_y_monotonic_range<F>(&self, cb: &mut F)
where F: FnMut(Range<S>),

Invokes a callback for each y-monotonic part of the segment.

pub fn for_each_y_monotonic<F>(&self, cb: &mut F)
where F: FnMut(&QuadraticBezierSegment<S>),

Invokes a callback for each y-monotonic part of the segment.

pub fn for_each_x_monotonic_range<F>(&self, cb: &mut F)
where F: FnMut(Range<S>),

Invokes a callback for each x-monotonic part of the segment.

pub fn for_each_x_monotonic<F>(&self, cb: &mut F)
where F: FnMut(&QuadraticBezierSegment<S>),

Invokes a callback for each x-monotonic part of the segment.

pub fn bounding_triangle(&self) -> Triangle<S>

Returns a triangle containing this curve segment.

pub fn fast_bounding_box(&self) -> Box2D<S>

Returns a conservative rectangle that contains the curve.

pub fn fast_bounding_range_x(&self) -> (S, S)

Returns a conservative range of x that contains this curve.

pub fn fast_bounding_range_y(&self) -> (S, S)

Returns a conservative range of y that contains this curve.

pub fn bounding_box(&self) -> Box2D<S>

Returns the smallest rectangle the curve is contained in

pub fn bounding_range_x(&self) -> (S, S)

Returns the smallest range of x that contains this curve.

pub fn bounding_range_y(&self) -> (S, S)

Returns the smallest range of y that contains this curve.

pub fn is_x_monotonic(&self) -> bool

Returns whether this segment is monotonic on the x axis.

pub fn is_y_monotonic(&self) -> bool

Returns whether this segment is monotonic on the y axis.

pub fn is_monotonic(&self) -> bool

Returns whether this segment is fully monotonic.

pub fn line_intersections_t(&self, line: &Line<S>) -> ArrayVec<S, 2>

Computes the intersections (if any) between this segment a line.

The result is provided in the form of the t parameters of each point along curve. To get the intersection points, sample the curve at the corresponding values.

pub fn line_intersections(&self, line: &Line<S>) -> ArrayVec<Point<S>, 2>

Computes the intersection points (if any) between this segment a line.

pub fn line_segment_intersections_t( &self, segment: &LineSegment<S>, ) -> ArrayVec<(S, S), 2>

Computes the intersections (if any) between this segment and a line segment.

The result is provided in the form of the t parameters of each point along curve and segment. To get the intersection points, sample the segments at the corresponding values.

pub fn from(&self) -> Point<S>

pub fn to(&self) -> Point<S>

pub fn line_segment_intersections( &self, segment: &LineSegment<S>, ) -> ArrayVec<Point<S>, 2>

Computes the intersection points (if any) between this segment a line segment.

pub fn closest_point(&self, pos: Point<S>) -> S

Analytic solution to finding the closest point on the curve to pos.

pub fn distance_to_point(&self, pos: Point<S>) -> S

Returns the shortest distance between this segment and a point.

pub fn square_distance_to_point(&self, pos: Point<S>) -> S

Returns the shortest squared distance between this segment and a point.

May be useful to avoid the cost of a square root when comparing against a distance that can be squared instead.

pub fn drag(&self, t: S, new_position: Point<S>) -> Self

pub fn length(&self) -> S

Computes the length of this segment.

Implements Raph Levien’s analytical approach described in https://raphlinus.github.io/curves/2018/12/28/bezier-arclength.html

pub fn to_f32(&self) -> QuadraticBezierSegment<f32>

pub fn to_f64(&self) -> QuadraticBezierSegment<f64>