Trait lyon_geom::segment::BoundingRect

pub trait BoundingRect {
    type Scalar: Scalar;
    fn bounding_rect(&self) -> Rect<Self::Scalar>;

    fn bounding_range_x(&self) -> (Self::Scalar, Self::Scalar);

    fn bounding_range_y(&self) -> (Self::Scalar, Self::Scalar);

    fn fast_bounding_range_x(&self) -> (Self::Scalar, Self::Scalar);

    fn fast_bounding_range_y(&self) -> (Self::Scalar, Self::Scalar);

    fn fast_bounding_rect(&self) -> Rect<Self::Scalar> { ... }
}

Associated Types

type Scalar: Scalar

Loading content...
Required methods

fn bounding_rect(&self) -> Rect<Self::Scalar>

Returns a rectangle that contains the curve.

fn bounding_range_x(&self) -> (Self::Scalar, Self::Scalar)

Returns a range of x values that contains the curve.

fn bounding_range_y(&self) -> (Self::Scalar, Self::Scalar)

Returns a range of y values that contains the curve.

fn fast_bounding_range_x(&self) -> (Self::Scalar, Self::Scalar)

Returns a range of x values that contains the curve.

fn fast_bounding_range_y(&self) -> (Self::Scalar, Self::Scalar)

Returns a range of y values that contains the curve.

Loading content...
Provided methods

fn fast_bounding_rect(&self) -> Rect<Self::Scalar>

Returns a rectangle that contains the curve.

This does not necessarily return the smallest possible bounding rectangle.

Loading content...