Struct lyon_geom::Angle

pub struct Angle<T> {
    pub radians: T,
}

An angle in radians. An angle in radians

Fields
radians: T
Implementations

impl<T> Angle<T>

pub fn radians(radians: T) -> Angle<T>

pub fn get(self) -> T

impl<T> Angle<T>where T: Trig,

pub fn degrees(deg: T) -> Angle<T>

pub fn to_degrees(self) -> T

impl<T> Angle<T>where T: Sub<T, Output = T> + Rem<T, Output = T> + Add<T, Output = T> + Zero + FloatConst + PartialOrd<T> + Copy,

pub fn positive(&self) -> Angle<T>

Returns this angle in the [0..2*PI[ range.

pub fn signed(&self) -> Angle<T>

Returns this angle in the ]-PI..PI] range.

impl<T> Angle<T>where T: Rem<T, Output = T> + Sub<T, Output = T> + Add<T, Output = T> + Mul<T, Output = T> + One + FloatConst + Copy,

pub fn angle_to(&self, to: Angle<T>) -> Angle<T>

Returns the shortest signed angle between two angles.

Takes wrapping and signs into account.

pub fn lerp(&self, other: Angle<T>, t: T) -> Angle<T>

Linear interpolation between two angles, using the shortest path.

impl<T> Angle<T>where T: Float,

pub fn is_finite(self) -> bool

Returns true if the angle is a finite number.

impl<T> Angle<T>where T: Real,

pub fn sin_cos(self) -> (T, T)

Returns (sin(self), cos(self)).

impl<T> Angle<T>where T: Zero,

pub fn zero() -> Angle<T>

impl<T> Angle<T>where T: FloatConst + Add<T, Output = T>,

pub fn pi() -> Angle<T>

pub fn two_pi() -> Angle<T>

pub fn frac_pi_2() -> Angle<T>

pub fn frac_pi_3() -> Angle<T>

pub fn frac_pi_4() -> Angle<T>

impl<T> Angle<T>where T: NumCast + Copy,

pub fn cast<NewT>(&self) -> Angle<NewT>where NewT: NumCast,

Cast from one numeric representation to another.

pub fn try_cast<NewT>(&self) -> Option<Angle<NewT>>where NewT: NumCast,

Fallible cast from one numeric representation to another.

pub fn to_f32(&self) -> Angle<f32>

Cast angle to f32.

pub fn to_f64(&self) -> Angle<f64>

Cast angle f64.