Struct lyon_geom::Arc

pub struct Arc<S> {
    pub center: Point<S>,
    pub radii: Vector<S>,
    pub start_angle: Angle<S>,
    pub sweep_angle: Angle<S>,
    pub x_rotation: Angle<S>,
}

An elliptic arc curve segment.

Fields
center: Point<S>radii: Vector<S>start_angle: Angle<S>sweep_angle: Angle<S>x_rotation: Angle<S>
Methods

impl<S: Scalar> Arc<S>

pub fn circle(center: Point<S>, radius: S) -> Self

Create simple circle.

pub fn from_svg_arc(arc: &SvgArc<S>) -> Arc<S>

Convert from the SVG arc notation.

pub fn to_svg_arc(&self) -> SvgArc<S>

Convert to the SVG arc notation.

pub fn for_each_quadratic_bezier<F>(&self, cb: &mut F) where
F: FnMut(&QuadraticBezierSegment<S>),

Approximate the arc with a sequence of quadratic bézier curves.

pub fn for_each_cubic_bezier<F>(&self, cb: &mut F) where
F: FnMut(&CubicBezierSegment<S>),

Approximate the arc with a sequence of cubic bézier curves.

pub fn sample(&self, t: S) -> Point<S>

Sample the curve at t (expecting t between 0 and 1).

pub fn x(&self, t: S) -> S

pub fn y(&self, t: S) -> S

pub fn sample_tangent(&self, t: S) -> Vector<S>

Sample the curve's tangent at t (expecting t between 0 and 1).

pub fn get_angle(&self, t: S) -> Angle<S>

Sample the curve's angle at t (expecting t between 0 and 1).

pub fn end_angle(&self) -> Angle<S>

pub fn from(&self) -> Point<S>

pub fn to(&self) -> Point<S>

pub fn split_range(&self, t_range: Range<S>) -> Self

Return the sub-curve inside a given range of t.

This is equivalent splitting at the range's end points.

pub fn split(&self, t: S) -> (Arc<S>, Arc<S>)

Split this curve into two sub-curves.

pub fn before_split(&self, t: S) -> Arc<S>

Return the curve before the split point.

pub fn after_split(&self, t: S) -> Arc<S>

Return the curve after the split point.

pub fn flip(&self) -> Self

Swap the direction of the segment.

pub fn for_each_flattened<F: FnMut(Point<S>)>(
&self,
tolerance: S,
call_back: &mut F
)

Approximates the arc with a sequence of line segments.

pub fn flattening_step(&self, tolerance: S) -> S

Finds the interval of the beginning of the curve that can be approximated with a line segment.

pub fn flattened(&self, tolerance: S) -> Flattened<S>

Returns the flattened representation of the curve as an iterator, starting after the current point.

pub fn fast_bounding_rect(&self) -> Rect<S>

Returns a conservative rectangle that contains the curve.

pub fn bounding_rect(&self) -> Rect<S>

Returns a conservative rectangle that contains the curve.

pub fn for_each_local_x_extremum_t<F>(&self, cb: &mut F) where
F: FnMut(S),

pub fn for_each_local_y_extremum_t<F>(&self, cb: &mut F) where
F: FnMut(S),

pub fn bounding_range_x(&self) -> (S, S)

pub fn bounding_range_y(&self) -> (S, S)

pub fn fast_bounding_range_x(&self) -> (S, S)

pub fn fast_bounding_range_y(&self) -> (S, S)

pub fn approximate_length(&self, tolerance: S) -> S