Struct lyon_geom::LineSegment

pub struct LineSegment<S> {
    pub from: Point<S>,
    pub to: Point<S>,
}

A linear segment.

Fields
from: Point<S>to: Point<S>
Methods

impl<S: Scalar> LineSegment<S>

pub fn sample(&self, t: S) -> Point<S>

Sample the segment at t (expecting t between 0 and 1).

pub fn x(&self, t: S) -> S

Sample the x coordinate of the segment at t (expecting t between 0 and 1).

pub fn y(&self, t: S) -> S

Sample the y coordinate of the segment at t (expecting t between 0 and 1).

pub fn from(&self) -> Point<S>

pub fn to(&self) -> Point<S>

pub fn solve_t_for_x(&self, x: S) -> S

pub fn solve_t_for_y(&self, y: S) -> S

pub fn solve_y_for_x(&self, x: S) -> S

pub fn solve_x_for_y(&self, y: S) -> S

pub fn flip(&self) -> Self

Returns an inverted version of this segment where the beginning and the end points are swapped.

pub fn split_range(&self, t_range: Range<S>) -> Self

Return the sub-segment inside a given range of t.

This is equivalent splitting at the range's end points.

pub fn split(&self, t: S) -> (Self, Self)

Split this curve into two sub-segments.

pub fn before_split(&self, t: S) -> Self

Return the segment before the split point.

pub fn after_split(&self, t: S) -> Self

Return the segment after the split point.

pub fn split_at_x(&self, x: S) -> (Self, Self)

pub fn bounding_rect(&self) -> Rect<S>

Return the minimum bounding rectangle

fn bounding_range_x(&self) -> (S, S)

fn bounding_range_y(&self) -> (S, S)

pub fn to_vector(&self) -> Vector<S>

Returns the vector between this segment's from and to points.

pub fn to_line(&self) -> Line<S>

Returns the line containing this segment.

pub fn length(&self) -> S

Computes the length of this segment.

pub fn set_length(&mut self, new_length: S)

Changes the segment's length, moving destination point.

pub fn translate(&mut self, by: Vector<S>) -> Self

pub fn transformed<T: Transformation<S>>(&self, transform: &T) -> Self

Applies the transform to this segment and returns the results.

pub fn intersection_t(&self, other: &Self) -> Option<(S, S)>

Computes the intersection (if any) between this segment and another one.

The result is provided in the form of the t parameter of each segment. To get the intersection point, sample one of the segments at the corresponding value.

pub fn intersection(&self, other: &Self) -> Option<Point<S>>

pub fn line_intersection_t(&self, line: &Line<S>) -> Option<S>

pub fn line_intersection(&self, line: &Line<S>) -> Option<Point<S>>

pub fn horizontal_line_intersection_t(&self, y: S) -> Option<S>

pub fn vertical_line_intersection_t(&self, x: S) -> Option<S>

pub fn horizontal_line_intersection(&self, y: S) -> Option<Point<S>>

pub fn vertical_line_intersection(&self, x: S) -> Option<Point<S>>

fn axis_aligned_intersection_1d(a: S, b: S, v: S) -> Option<S>

pub fn intersects(&self, other: &Self) -> bool

pub fn intersects_line(&self, line: &Line<S>) -> bool

pub fn overlaps_line(&self, line: &Line<S>) -> bool

pub fn overlaps_segment(&self, other: &LineSegment<S>) -> bool

pub fn contains_segment(&self, other: &LineSegment<S>) -> bool