Struct lyon_geom::QuadraticBezierSegment

pub struct QuadraticBezierSegment<S> {
    pub from: Point<S>,
    pub ctrl: Point<S>,
    pub to: Point<S>,
}

A 2d curve segment defined by three points: the beginning of the segment, a control point and the end of the segment.

The curve is defined by equation: ∀ t ∈ [0..1], P(t) = (1 - t)² * from + 2 * (1 - t) * t * ctrl + 2 * t² * to

Fields
from: Point<S> ctrl: Point<S> to: Point<S>
Methods

impl<S: Scalar> QuadraticBezierSegment<S>

pub fn sample(&self, t: S) -> Point<S>

Sample the curve at t (expecting t between 0 and 1).

pub fn x(&self, t: S) -> S

Sample the x coordinate of the curve at t (expecting t between 0 and 1).

pub fn y(&self, t: S) -> S

Sample the y coordinate of the curve at t (expecting t between 0 and 1).

pub fn derivative(&self, t: S) -> Vector<S>

Sample the curve's derivative at t (expecting t between 0 and 1).

pub fn dx(&self, t: S) -> S

Sample the x coordinate of the curve's derivative at t (expecting t between 0 and 1).

pub fn dy(&self, t: S) -> S

Sample the y coordinate of the curve's derivative at t (expecting t between 0 and 1).

pub fn flip(&self) -> Self

Swap the beginning and the end of the segment.

pub fn y_maximum_t(&self) -> S

Find the advancement of the y-most position in the curve.

This returns the advancement along the curve, not the actual y position.

pub fn y_minimum_t(&self) -> S

Find the advancement of the y-least position in the curve.

This returns the advancement along the curve, not the actual y position.

pub fn local_y_extremum_t(&self) -> Option<S>

Return the y inflection point or None if this curve is y-monotonic.

pub fn x_maximum_t(&self) -> S

Find the advancement of the x-most position in the curve.

This returns the advancement along the curve, not the actual x position.

pub fn x_minimum_t(&self) -> S

Find the advancement of the x-least position in the curve.

This returns the advancement along the curve, not the actual x position.

pub fn local_x_extremum_t(&self) -> Option<S>

Return the x inflection point or None if this curve is x-monotonic.

pub fn split_range(&self, t_range: Range<S>) -> Self

Return the sub-curve inside a given range of t.

This is equivalent splitting at the range's end points.

pub fn split(
&self,
t: S
) -> (QuadraticBezierSegment<S>, QuadraticBezierSegment<S>)

Split this curve into two sub-curves.

pub fn before_split(&self, t: S) -> QuadraticBezierSegment<S>

Return the curve before the split point.

pub fn after_split(&self, t: S) -> QuadraticBezierSegment<S>

Return the curve after the split point.

pub fn to_cubic(&self) -> CubicBezierSegment<S>

Elevate this curve to a third order bézier.

pub fn baseline(&self) -> LineSegment<S>

pub fn is_linear(&self, tolerance: S) -> bool

pub fn fat_line(&self) -> (LineEquation<S>, LineEquation<S>)

Computes a "fat line" of this segment.

A fat line is two convervative lines between which the segment is fully contained.

pub fn transform(&self, transform: &Transform2D<S>) -> Self

Applies the transform to this curve and returns the results.

pub fn flattening_step(&self, tolerance: S) -> S

Find the interval of the begining of the curve that can be approximated with a line segment.

pub fn for_each_flattened<F: FnMut(Point<S>)>(
&self,
tolerance: S,
call_back: &mut F
)

Iterates through the curve invoking a callback at each point.

pub fn flattened(&self, tolerance: S) -> Flattened<S>

Returns the flattened representation of the curve as an iterator, starting after the current point.

pub fn for_each_monotonic_t<F>(&self, cb: F) where
F: FnMut(S),

Invokes a callback between each monotonic part of the segment.

pub fn for_each_monotonic_range<F>(&self, cb: F) where
F: FnMut(Range<S>),

Invokes a callback for each monotonic part of the segment..

pub fn for_each_monotonic<F>(&self, cb: &mut F) where
F: FnMut(&Monotonic<QuadraticBezierSegment<S>>),

pub fn approximate_length(&self, tolerance: S) -> S

Compute the length of the segment using a flattened approximation.

pub fn bounding_triangle(&self) -> Triangle<S>

Returns a triangle containing this curve segment.

pub fn fast_bounding_rect(&self) -> Rect<S>

Returns a conservative rectangle that contains the curve.

pub fn fast_bounding_range_x(&self) -> (S, S)

Returns a conservative range of x this curve is contained in.

pub fn fast_bounding_range_y(&self) -> (S, S)

Returns a conservative range of y this curve is contained in.

pub fn bounding_rect(&self) -> Rect<S>

Returns the smallest rectangle the curve is contained in

pub fn bounding_range_x(&self) -> (S, S)

Returns the smallest range of x this curve is contained in.

pub fn bounding_range_y(&self) -> (S, S)

Returns the smallest range of y this curve is contained in.

pub fn assume_monotonic(&self) -> MonotonicQuadraticBezierSegment<S>

Cast this curve into a monotonic curve without checking that the monotonicity assumption is correct.

pub fn is_x_monotonic(&self) -> bool

Returns whether this segment is monotonic on the x axis.

pub fn is_y_monotonic(&self) -> bool

Returns whether this segment is monotonic on the y axis.

pub fn is_monotonic(&self) -> bool

Returns whether this segment is fully monotonic.

pub fn line_intersections_t(&self, line: &Line<S>) -> ArrayVec<[S; 2]>

Computes the intersections (if any) between this segment a line.

The result is provided in the form of the t parameters of each point along curve. To get the intersection points, sample the curve at the corresponding values.

pub fn line_intersections(&self, line: &Line<S>) -> ArrayVec<[Point<S>; 2]>

Computes the intersection points (if any) between this segment a line.

pub fn line_segment_intersections_t(
&self,
segment: &LineSegment<S>
) -> ArrayVec<[(S, S); 2]>

Computes the intersections (if any) between this segment a line segment.

The result is provided in the form of the t parameters of each point along curve and segment. To get the intersection points, sample the segments at the corresponding values.

pub fn from(&self) -> Point<S>

pub fn to(&self) -> Point<S>

pub fn line_segment_intersections(
&self,
segment: &LineSegment<S>
) -> ArrayVec<[Point<S>; 2]>

Computes the intersection points (if any) between this segment a line segment.