Struct lyon_geom::SvgArc

pub struct SvgArc<S> {
    pub from: Point<S>,
    pub to: Point<S>,
    pub radii: Vector<S>,
    pub x_rotation: Angle<S>,
    pub flags: ArcFlags,
}

An elliptic arc curve segment using the SVG's end-point notation.

Fields
from: Point<S> to: Point<S> radii: Vector<S> x_rotation: Angle<S> flags: ArcFlags
Methods

impl<S: Scalar> SvgArc<S>

pub fn to_arc(&self) -> Arc<S>

Converts this arc from endpoints to center notation.

pub fn is_straight_line(&self) -> bool

Per SVG spec, this arc should be rendered as a line_to segment.

Do not convert an SvgArc into an arc if this returns true.

pub fn for_each_quadratic_bezier<F>(&self, cb: &mut F) where
F: FnMut(&QuadraticBezierSegment<S>),

Approximates the arc with a sequence of quadratic bézier segments.

pub fn for_each_cubic_bezier<F>(&self, cb: &mut F) where
F: FnMut(&CubicBezierSegment<S>),

Approximates the arc with a sequence of cubic bézier segments.

pub fn for_each_flattened<F: FnMut(Point<S>)>(&self, tolerance: S, cb: &mut F)

Approximates the arc with a sequence of line segments.