Trait lyon_geom::traits::FlatteningStep

pub trait FlatteningStep: Segment {
    fn flattening_step(&self, tolerance: Self::Scalar) -> Self::Scalar;
}

Types that implement local flattening approximation at the start of the curve.

Required methods

fn flattening_step(&self, tolerance: Self::Scalar) -> Self::Scalar

Find the interval of the begining of the curve that can be approximated with a line segment.