Struct lyon_geom::triangle::Triangle

pub struct Triangle<S> {
    pub a: Point<S>,
    pub b: Point<S>,
    pub c: Point<S>,
}

A 2D triangle defined by three points a, b and c.

Fields
a: Point<S>b: Point<S>c: Point<S>
Implementations

impl<S: Scalar> Triangle<S>

fn get_barycentric_coords_for_point(&self, point: Point<S>) -> (S, S, S)

pub fn contains_point(&self, point: Point<S>) -> bool

pub fn bounding_rect(&self) -> Rect<S>

Return the minimum bounding rectangle.

pub fn ab(&self) -> LineSegment<S>

pub fn ba(&self) -> LineSegment<S>

pub fn bc(&self) -> LineSegment<S>

pub fn cb(&self) -> LineSegment<S>

pub fn ca(&self) -> LineSegment<S>

pub fn ac(&self) -> LineSegment<S>

pub fn transform<T: Transformation<S>>(&self, transform: &T) -> Self

[Not implemented] Applies the transform to this triangle and returns the results.

pub fn intersects(&self, other: &Self) -> bool

Test for triangle-triangle intersection.

pub fn intersects_line_segment(&self, segment: &LineSegment<S>) -> bool

Test for triangle-segment intersection.