Trait lyon_geom::Segment

pub trait Segment: Copy + Sized {
    type Scalar: Scalar;
    pub fn from(&self) -> Point<Self::Scalar>;

    pub fn to(&self) -> Point<Self::Scalar>;

    pub fn sample(&self, t: Self::Scalar) -> Point<Self::Scalar>;

    pub fn derivative(&self, t: Self::Scalar) -> Vector<Self::Scalar>;

    pub fn split(&self, t: Self::Scalar) -> (Self, Self);

    pub fn before_split(&self, t: Self::Scalar) -> Self;

    pub fn after_split(&self, t: Self::Scalar) -> Self;

    pub fn split_range(&self, t_range: Range<Self::Scalar>) -> Self;

    pub fn flip(&self) -> Self;

    pub fn approximate_length(&self, tolerance: Self::Scalar) -> Self::Scalar;

    pub fn x(&self, t: Self::Scalar) -> Self::Scalar { ... }

    pub fn y(&self, t: Self::Scalar) -> Self::Scalar { ... }

    pub fn dx(&self, t: Self::Scalar) -> Self::Scalar { ... }

    pub fn dy(&self, t: Self::Scalar) -> Self::Scalar { ... }
}

Common APIs to segment types.

Associated Types

type Scalar: Scalar

Loading content...
Required methods

pub fn from(&self) -> Point<Self::Scalar>

Start of the curve.

pub fn to(&self) -> Point<Self::Scalar>

End of the curve.

pub fn sample(&self, t: Self::Scalar) -> Point<Self::Scalar>

Sample the curve at t (expecting t between 0 and 1).

pub fn derivative(&self, t: Self::Scalar) -> Vector<Self::Scalar>

Sample the derivative at t (expecting t between 0 and 1).

pub fn split(&self, t: Self::Scalar) -> (Self, Self)

Split this curve into two sub-curves.

pub fn before_split(&self, t: Self::Scalar) -> Self

Return the curve before the split point.

pub fn after_split(&self, t: Self::Scalar) -> Self

Return the curve after the split point.

pub fn split_range(&self, t_range: Range<Self::Scalar>) -> Self

Return the curve inside a given range of t.

This is equivalent splitting at the range's end points.

pub fn flip(&self) -> Self

Swap the direction of the segment.

pub fn approximate_length(&self, tolerance: Self::Scalar) -> Self::Scalar

Compute the length of the segment using a flattened approximation.

Loading content...
Provided methods

pub fn x(&self, t: Self::Scalar) -> Self::Scalar

Sample x at t (expecting t between 0 and 1).

pub fn y(&self, t: Self::Scalar) -> Self::Scalar

Sample y at t (expecting t between 0 and 1).

pub fn dx(&self, t: Self::Scalar) -> Self::Scalar

Sample x derivative at t (expecting t between 0 and 1).

pub fn dy(&self, t: Self::Scalar) -> Self::Scalar

Sample y derivative at t (expecting t between 0 and 1).

Loading content...