Trait lyon_geom::traits::FlatteningStep

pub trait FlatteningStep: FlattenedForEach {
    fn flattening_step(&self, tolerance: Self::Scalar) -> Self::Scalar;

    fn flattened(self, tolerance: Self::Scalar) -> Flattened<Self::Scalar, Self> { ... }
}

Types that implement local flattening approximation at the start of the curve.

Required Methods

fn flattening_step(&self, tolerance: Self::Scalar) -> Self::Scalar

Find the interval of the begining of the curve that can be approximated with a line segment.

Provided Methods

fn flattened(self, tolerance: Self::Scalar) -> Flattened<Self::Scalar, Self>

Returns the flattened representation of the curve as an iterator, starting after the current point.